+++ library/Kaltura/Client/Type/StatsKmcEvent.php
@@ -58,7 +58,7 @@ class StatsKmcEvent extends \Kaltura\Client\ObjectBase
 		if(count($xml->kmcEventType))
 			$this->kmcEventType = (int)$xml->kmcEventType;
 		if(count($xml->eventTimestamp))
-			$this->eventTimestamp = (float)$xml->eventTimestamp;
+			$this->eventTimestamp = (int)$xml->eventTimestamp;
 		if(count($xml->sessionId))
 			$this->sessionId = (string)$xml->sessionId;
 		if(count($xml->partnerId))
@@ -94,7 +94,7 @@ class StatsKmcEvent extends \Kaltura\Client\ObjectBase
 
 	/**
 	 * the client's timestamp of this event
-	 * @var float
+	 * @var int
 	 */
 	public $eventTimestamp = null;
